Todos/pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.115.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "pytest-asyncio>=0.25.2",
    "pytest>=8.3.4",
    "sqlalchemy>=2.0.37",
]
